Add files via upload
This HTML file presents the final output of the DineLex project, a restaurant jargon lexicon. The document visually summarizes the distribution of categories within the lexicon, providing insight into various jargon terms related to dining experiences. The data visualization in the form of a pie chart illustrates the proportion of terms categorized under Ambiance, Drinks, Food, Service, and Other. The file has been knitted from an R Markdown file and is hosted on GitHub Pages, allowing for easy sharing and access.